Factor out date-util.ts.
