+++ frontend/lib/loc/access-dates.tsx
@@ -9,7 +9,7 @@ import {
 } from "../queries/AccessDatesMutation";
 import { AccessDatesInput } from "../queries/globalTypes";
 import { ProgressButtons } from "../ui/buttons";
-import { dateAsISO, addDays } from "../util/util";
+import { dateAsISO, addDays } from "../util/date-util";
 
 import validation from "../../../common-data/access-dates-validation.json";
 import { MiddleProgressStep } from "../progress/progress-step-route";
