+++ frontend/lib/util/tests/util.test.ts
@@ -1,8 +1,5 @@
 import {
   assertNotNull,
-  dateAsISO,
-  addDays,
-  friendlyDate,
   callOnceWithinMs,
   getFunctionProperty,
   exactSubsetOrDefault,
@@ -50,44 +47,6 @@ describe("assertNotUndefined()", () => {
 
   it("returns argument when not undefined", () => {
     expect(assertNotUndefined(null)).toBe(null);
-  });
-});
-
-test("dateAsISIO() works", () => {
-  expect(dateAsISO(new Date(2018, 0, 2))).toBe("2018-01-02");
-});
-
-test("addDays() works", () => {
-  expect(addDays(new Date(2018, 0, 30), 7).toDateString()).toBe(
-    "Tue Feb 06 2018"
-  );
-});
-
-describe("friendlyDate()", () => {
-  const dateStr = "2018-01-02T04:00:00.000Z";
-
-  it("translates to time zone on platforms that support it", () => {
-    expect(friendlyDate(new Date(dateStr), "America/New_York")).toBe(
-      "Monday, January 1, 2018"
-    );
-  });
-
-  it("falls back to local timezone if platform sucks", () => {
-    jest.spyOn(Intl, "DateTimeFormat").mockImplementationOnce(() => {
-      throw new Error("i am ie11 and idk what a time zone is");
-    });
-    expect(friendlyDate(new Date(dateStr), "America/New_York")).toMatch(
-      /January/
-    );
-  });
-
-  it("falls back to janky string if platform really sucks", () => {
-    jest.spyOn(Intl, "DateTimeFormat").mockImplementation(() => {
-      throw new Error("i am a really old browser and idk what Intl is");
-    });
-    expect(friendlyDate(new Date(dateStr), "America/New_York")).toMatch(
-      /Jan 0/
-    );
   });
 });
